cleaned up adapters
