+++ packages/adapters/README.md
@@ -1,4 +1,3 @@
 # Adaprters
 
 To convert STT json transcript into draft.js code block.
-Used by TimedTextEditor.
